indexerSlice cleanup
