+++ src/features/indexer/indexerRegistryApi.ts
@@ -1,14 +1,6 @@
-import {
-  IndexedOrderResponse,
-  IndexerRegistry,
-  NodeIndexer,
-  RequestFilter,
-} from "@airswap/libraries";
-import { FullOrderERC20 } from "@airswap/types";
+import { IndexerRegistry } from "@airswap/libraries";
 
 import { providers } from "ethers";
-
-import { INDEXER_ORDER_RESPONSE_TIME_MS } from "../../constants/configParams";
 
 export const getIndexerUrls = async (
   chainId: number,
@@ -16,29 +8,4 @@ export const getIndexerUrls = async (
 ): Promise<string[]> => {
   const indexerRegistry = new IndexerRegistry(chainId, provider);
   return await indexerRegistry.contract.getURLs();
-};
-
-export const getIndexerOrders = async (
-  indexers: NodeIndexer[] | undefined,
-  filter: RequestFilter
-): Promise<FullOrderERC20[]> => {
-  let orders: Record<string, IndexedOrderResponse> = {};
-
-  const orderPromises = indexers?.map(async (indexer, i) => {
-    try {
-      const orderResponse = await indexer.getOrdersERC20By(filter);
-      const ordersToAdd = orderResponse.orders;
-      orders = { ...orders, ...ordersToAdd };
-    } catch (e) {
-      console.log(
-        `[indexerSlice] Order request failed for ${indexers[i]}`,
-        e || ""
-      );
-    }
-  });
-
-  return Promise.race([
-    orderPromises && Promise.allSettled(orderPromises),
-    new Promise((res) => setTimeout(res, INDEXER_ORDER_RESPONSE_TIME_MS)),
-  ]).then(() => Object.entries(orders).map(([, order]) => order.order));
 };
