+++ src/features/indexer/indexerSlice.ts
@@ -1,4 +1,5 @@
 import {
+  IndexedOrderResponse,
   NodeIndexer,
   RequestFilter,
   SortField,
@@ -10,22 +11,23 @@ import { createAsyncThunk, createSlice } from "@reduxjs/toolkit";
 import { providers } from "ethers";
 
 import { AppDispatch, RootState } from "../../app/store";
-import i18n from "../../i18n/i18n";
+import { INDEXER_ORDER_RESPONSE_TIME_MS } from "../../constants/configParams";
 import { getIndexerOrders, getIndexerUrls } from "./indexerRegistryApi";
 
 export interface IndexerState {
   /** List of indexer urls for servers that have responded successfully to
    * the healthcheck within the allowed time. Null during initial fetch. */
   indexerUrls: string[] | null;
-  /** Map of order hash -> fullorder */
   orders: FullOrderERC20[];
-  helperText: string | null;
+  isLoading: boolean;
+  noIndexersFound: boolean;
 }
 
 const initialState: IndexerState = {
   indexerUrls: null,
   orders: [],
-  helperText: null,
+  isLoading: true,
+  noIndexersFound: false,
 };
 
 export const fetchIndexerUrls = createAsyncThunk<
@@ -48,15 +50,29 @@ export const getFilteredOrders = createAsyncThunk<
 >("indexer/getFilteredOrders", async ({ filter }, { getState }) => {
   const { indexer: indexerState } = getState();
 
-  const indexers = indexerState.indexerUrls?.map(
-    (indexer) => new NodeIndexer(indexer)
-  );
+  const indexers = indexerState.indexerUrls?.map((url) => new NodeIndexer(url));
 
-  return await getIndexerOrders(indexers, {
-    ...filter,
-    sortField: SortField.SIGNER_AMOUNT,
-    sortOrder: SortOrder.DESC,
+  let orders: Record<string, IndexedOrderResponse> = {};
+
+  const orderPromises = indexers?.map(async (indexer, i) => {
+    try {
+      const orderResponse = await indexer.getOrdersERC20By(filter);
+      const ordersToAdd = orderResponse.orders;
+      orders = { ...orders, ...ordersToAdd };
+    } catch (e) {
+      console.log(
+        `[indexerSlice] Order request failed for ${
+          indexerState.indexerUrls![i] || "an indexer node"
+        }`,
+        e || ""
+      );
+    }
   });
+
+  return Promise.race([
+    orderPromises && Promise.allSettled(orderPromises),
+    new Promise((res) => setTimeout(res, INDEXER_ORDER_RESPONSE_TIME_MS)),
+  ]).then(() => Object.entries(orders).map(([, order]) => order.order));
 });
 
 export const indexerSlice = createSlice({
@@ -75,20 +91,16 @@ export const indexerSlice = createSlice({
 
       state.indexerUrls = action.payload;
 
-      if (!action.payload.length) {
-        state.helperText = i18n.t("orders.noIndexersFound");
+      if (!state.indexerUrls.length) {
+        state.noIndexersFound = true;
       }
     });
     builder.addCase(getFilteredOrders.fulfilled, (state, action) => {
       state.orders = action.payload;
-      if (!action.payload.length) {
-        state.helperText = i18n.t("orders.noIndexerOrdersFound");
-      }
+      state.isLoading = false;
     });
     builder.addCase(getFilteredOrders.rejected, (state) => {
-      if (!state.orders.length) {
-        state.helperText = i18n.t("orders.noIndexerOrdersFound");
-      }
+      state.isLoading = false;
     });
   },
 });
